// src/types/types.ts
export interface ColumnContainerProps {
    color: string;
}

export interface ColumnHeaderProps {
    color: string;
}

export interface ColumnTitleProps {
    color: string;
}

export interface AddCardButtonProps {
    color: string;
}

export interface CountBadgeProps {
    color: string;
}

export interface AddTaskCardPropsStyles {
    color: string;
}